+++ includes/styles/spacing.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.4.0
+ * @version  1.7.0
  */
 
 namespace WebManDesign\ABS;
@@ -17,8 +17,9 @@ defined( 'ABSPATH' ) || exit;
 Register::add_style(
 	'no-gap-vertical',
 	array(
-		'label'  => _x( 'No vertical gap', 'Block style label.', 'additional-block-styles' ),
-		'blocks' => array(
+		'label'       => _x( 'No vertical gap', 'Block style label.', 'additional-block-styles' ),
+		'description' => _x( 'Not needed if block margin options are supported.', 'Block style label.', 'additional-block-styles' ),
+		'blocks'      => array(
 			'core/column',
 			'core/columns',
 			'core/cover',
@@ -36,8 +37,9 @@ Register::add_style(
 Register::add_style(
 	'no-gaps',
 	array(
-		'label'  => _x( 'No gaps', 'Block style label.', 'additional-block-styles' ),
-		'blocks' => array(
+		'label'       => _x( 'No gaps', 'Block style label.', 'additional-block-styles' ),
+		'description' => _x( 'Not needed if block spacing options are supported.', 'Block style label.', 'additional-block-styles' ),
+		'blocks'      => array(
 			'core/columns',
 			'core/gallery',
 			'core/group',